Resolve SonarCloud Minor Bugs (#1449)
* fix few minor sonarcloud bugs


* Additional fixes


* fix code smell


* move duplicated code to a method


* fix other minor bugs


* fix other issues


* remove unnecessary toString() calls


* add braces


* remove a few casts


* convert getCommonTransactionBodyBytes and legacyReceiptStorageSecs to return long and remove casts


* fix other bugs created


* Address review comments and fix alignment in SmartContractFeeBuilder


* address few review comments


* fix few other review points


* fix typo


* fix review comments, add UnitTest for MerkleDiskFs


* Add unit tests for FreezeHandler


* fix random failing tests because of MockedStatic


* add more tests for codecov


* fix all failing tests, yet to add more tests for code coverage


* remove main method in SignatureVerifier and unused methods as they are not used anywhere


* Address other review comments

+++ hapi-fees/src/main/java/com/hedera/services/usage/schedule/ExtantScheduleContext.java
@@ -33,7 +33,7 @@ import static com.hederahashgraph.fee.FeeBuilder.BOOL_SIZE;
 import static com.hederahashgraph.fee.FeeBuilder.getAccountKeyStorageSize;
 
 public class ExtantScheduleContext {
-	static final int METADATA_SIZE =
+	static final long METADATA_SIZE =
 			/* The schedule id, the scheduling account, and the responsible payer */
 			3 * BASIC_ENTITY_ID_SIZE +
 					/* The expiration time */
